:recycle: Extract MotifElimination component
