+++ src/views/certificates/ppe2/components/motifElimination/DejaLaureatNonInstruit.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const DejaLaureatNonInstruit = () => (
+  <>
+    À la suite de l'instruction par les services de la Commission de régulation de l’énergie, je
+    suis au regret de vous informer que votre offre n’a pas été retenue car elle avait déjà été
+    désignée lauréate au cours d’un précédent appel d’offres.
+  </>
+)
