+++ src/views/certificates/ppe2/components/motifElimination/index.ts
@@ -1,0 +1,1 @@
+export * from './MotifElimination'
